done now fr fr fr

+++ backend/db/models/spot.js
@@ -26,7 +26,7 @@ module.exports = (sequelize, DataTypes) => {
     lng: DataTypes.DECIMAL,
     name: DataTypes.STRING,
     description: DataTypes.STRING,
-    price: DataTypes.DECIMAL
+    price: DataTypes.INTEGER
   }, {
     sequelize,
     modelName: 'Spot',
